OK the timers are now positioned correctly.
Need to get the platform ready timer to appear and start when the center referee presses the plaform ready button

also need the next attempt timer to only appear once the 3 referee decisions are given.

the reset button should affect the Platform ready timer but not the next attempt timer.